backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.0
openai==1.10.0
gunicorn==21.2.0
